Corrected mistake in double method signature on WriteStream.

+++ trunk/crinch/crinch-util/src/main/java/com/tomgibara/crinch/util/AbstractWriteStream.java
@@ -61,7 +61,7 @@ public abstract class AbstractWriteStream implements WriteStream {
 	}
 	
 	@Override
-	public void writeDouble(float v) {
+	public void writeDouble(double v) {
 		writeLong(Double.doubleToLongBits(v));
 	}
